Experimental ARM support
It can convert a TC, and it starts, but my raspberry pi isn't setup for
running graphical applications, so I can't verify this properly.

According to documentation ARM is bi-endian but usually uses
little-endian, and it's little-endian on my raspberry pi so let's
assume this will work until it doesn't.