ci: update CI
* Switch to vcpkg because it's a heck of a lot faster.
* Use well-defined vcpkg actions (caching, etc etc)
* Disable release action because it's a bit strange, e.g.:

It doesn't make use of pre-releases, which is what the current release
action is intended for.

A better solution would be to have two different "release" actions.

One that triggers on matching tags that are pushed to the repository
(e.g. v1.4.0) that builds & does an actual release.

Another that does "nightly" builds once a day that builds, releases a
pre-release tag (e.g. nightly-$DATE) and does not trigger when there
have been no commits since the last nightly.